changed return

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
     public double calculate (double bodyMassKg, double heightMetre) {
         double index = bodyMassKg / (heightMetre * heightMetre);
-        return index;
+        return (int)index;
     }
 }
